Refactored keygen interface
