+++ backend/app/usecase/keygen/keygen.go
@@ -1,8 +1,63 @@
 package keygen
 
-import "github.com/short-d/kgs/app/entity"
+import (
+	"errors"
 
-// KeyGenerator generates unique keys.
-type KeyGenerator interface {
-	NewKey() (entity.Key, error)
+	"github.com/short-d/kgs/app/entity"
+	"github.com/short-d/short/app/usecase/service"
+)
+
+type bufferEntry struct {
+	key entity.Key
+	err error
+}
+
+// KeyGenerator represents a remote KeyGenerator which fetch unique keys from remote service
+// and buffer them in memory for fast response.
+type KeyGenerator struct {
+	bufferSize int
+	buffer     chan bufferEntry
+	keyFetcher service.KeyFetcher
+}
+
+// NewKey produces a key
+func (r KeyGenerator) NewKey() (entity.Key, error) {
+	if len(r.buffer) == 0 {
+		go func() {
+			r.fetchKeys()
+		}()
+	}
+
+	entry := <-r.buffer
+	return entry.key, entry.err
+}
+
+func (r KeyGenerator) fetchKeys() {
+	keys, err := r.keyFetcher.FetchKeys(r.bufferSize)
+	if err != nil {
+		r.buffer <- bufferEntry{
+			key: "",
+			err: err,
+		}
+		return
+	}
+
+	for _, key := range keys {
+		r.buffer <- bufferEntry{
+			key: key,
+			err: nil,
+		}
+	}
+}
+
+// NewKeyGenerator creates Remote keygen generator
+func NewKeyGenerator(bufferSize int, keyFetcher service.KeyFetcher) (KeyGenerator, error) {
+	if bufferSize < 1 {
+		return KeyGenerator{}, errors.New("buffer size can't be less than 1")
+	}
+	return KeyGenerator{
+		bufferSize: bufferSize,
+		buffer:     make(chan bufferEntry, bufferSize),
+		keyFetcher: keyFetcher,
+	}, nil
 }
